+++ Sources/MapboxNavigation/MapboxSpeechSynthesizer.swift
@@ -54,10 +54,16 @@ open class MapboxSpeechSynthesizer: NSObject, SpeechSynthesizing {
     
     // MARK: Instructions vocalization
     
+    /// Checks if speech synthesizer is now pronouncing an instruction.
     public var isSpeaking: Bool {
         return audioPlayer?.isPlaying ?? false
     }
     
+    
+    /// Creates new `MapboxSpeechSynthesizer` with standard `SpeechSynthesizer` for converting text to audio.
+    ///
+    /// - parameter accessToken: A Mapbox [access token](https://www.mapbox.com/help/define-access-token/) used to authorize Mapbox Voice API requests. If an access token is not specified when initializing the speech synthesizer object, it should be specified in the `MBXAccessToken` key in the main application bundle’s Info.plist.
+    /// - parameter host: An optional hostname to the server API. The Mapbox Voice API endpoint is used by default.
     public init(accessToken: String? = nil, host: String? = nil) {
         self.cache = DataCache()
         
@@ -67,6 +73,14 @@ open class MapboxSpeechSynthesizer: NSObject, SpeechSynthesizing {
         }
         
         self.remoteSpeechSynthesizer = SpeechSynthesizer(accessToken: accessToken, host: hostString)
+    }
+    
+    /// Creates new `MapboxSpeechSynthesizer` with provided `SpeechSynthesizer` instance for converting text to audio.
+    ///
+    /// - parameter remoteSpeechSynthesizer: Custom `SpeechSynthesizer` used to provide voice data.
+    public init(remoteSpeechSynthesizer: SpeechSynthesizer) {
+        self.cache = DataCache()
+        self.remoteSpeechSynthesizer = remoteSpeechSynthesizer
     }
     
     deinit {